Changed variable type for MCP3426 to be integer

+++ include/Soil_sensor/MCP342X.h
@@ -5,7 +5,7 @@
     @author   C. Schnarel, I. Mitchell
     @license  BSD (see license.txt)
 
-    
+
     This is part of an Arduino library to interface with the Microchip
     MCP47X6 series of Analog-to-Digital converters which are connected
     via the I2C bus. Based on Microchip datasheets for the following part numbers
@@ -122,7 +122,7 @@ class MCP342X {
  	* @brief Test connection to the MCP3426X ADC.
  	*/
         bool testConnection(void);
-	
+
 	/*!
         * @brief Set the configuration bits for the ADC.
 	*/
@@ -138,13 +138,13 @@ class MCP342X {
 	* @brief Read the ADC result.
 	* @param[] data :
 	*/
-        uint8_t getResult(uint8_t *data);
+        int getResult(int *data);
 
         /*!
 	* @brief Non-blocking read the ADC result.
 	* @param[] data :
 	*/
-        uint8_t checkforResult(uint8_t *data);
+        int checkforResult(int *data);
 
     private:
 	/*!
